Rename transitionPage.ts from camelCase to kebab-case
