+++ src/server/inversify.config.ts
@@ -3,7 +3,7 @@ import { UrlCacheRedis } from './api/cache/url'
 import { UrlRepositorySequelize } from './api/repositories/url'
 import { GaLogger } from './api/analytics/analyticsLogger'
 import { DependencyIds } from './constants'
-import { CookieArrayReducer } from './util/transitionPage'
+import { CookieArrayReducer } from './util/transition-page'
 import { OtpCacheRedis } from './api/cache/otp'
 import { UserRepositorySequelize } from './api/repositories/user'
 import { MailerNode } from './util/email'
